requirements.txt
opencv-python
numpy
Pillow
pyautogui
pandas
openpyxl
easyocr
pytesseract
google-generativeai
pyyaml
requests
psutil